backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-aiplatform
google-auth
python-multipart
pydantic
